+++ scripts/macros/2024/classFeatures/multiple/evasion.js
@@ -1,0 +1,5 @@
+export let evasion = {
+    name: 'Evasion',
+    version: '1.3.36',
+    rules: 'modern'
+};